Fix 19 Eclipse warnings about unnecessary casts

+++ com.ibm.wala.cast.js/source/com/ibm/wala/cast/js/ipa/callgraph/JSCallGraphUtil.java
@@ -240,7 +240,7 @@ public class JSCallGraphUtil extends com.ibm.wala.cast.ipa.callgraph.CAstCallGra
       return (String)v;
     } else if (v instanceof Double) {
       String result = v.toString();
-      if (((double) Math.round((Double)v)) == ((Double)v).doubleValue()) {
+      if ((Math.round((Double)v)) == ((Double)v).doubleValue()) {
         result = Long.toString(Math.round((Double)v));
       }
       return result;
